backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
scipy
python-multipart
Flask
flask-cors
google-generativeai
